# app/Http/Controllers/RegisteredUserController.php
<?php

namespace App\Http\Controllers;

use App\Http\Controllers\Controller;
use App\Models\User;
use Illuminate\Http\Request;
use Illuminate\Support\Str;
use Symfony\Component\HttpFoundation\JsonResponse;
use App\Classes\ApiResponseClass;
use App\Services\GoogleServices\GoogleSheetsService;
use Illuminate\Support\Facades\Log;

use Illuminate\Support\Facades\Validator;

/**
 * @OA\Tag(name="Authentication")
 */
class RegisteredUserController extends Controller
{
    /**
     * @OA\Post(
     *     path="/api/authentication/validate-credentials",
     *     summary="Validate user credentials before registration",
     *     tags={"Authentication"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"email", "password"},
     *             @OA\Property(property="email", type="string", format="email", example="user@example.com"),
     *             @OA\Property(property="password", type="string", example="password123")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Credentials are valid",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=true)
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error"
     *     )
     * )
     * Handle an incoming registration request.
     *
     * @throws \Illuminate\Validation\ValidationException
     */
    public function validate(Request $request): JsonResponse
    {
        $validator = Validator::make($request->all(), User::rules(), User::messages());

        if ($validator->fails()) {
            return ApiResponseClass::sendInvalidFields($validator->errors()->toArray(), User::messages());
        }

        return ApiResponseClass::sendSuccess();
    }

    /**
     * @OA\Post(
     *     path="/api/authentication/register",
     *     summary="Register a new user",
     *     tags={"Authentication"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"email", "password", "name"},
     *             @OA\Property(property="email", type="string", format="email", example="user@example.com"),
     *             @OA\Property(property="password", type="string", example="password123"),
     *             @OA\Property(property="name", type="string", example="John Doe"),
     *             @OA\Property(property="phone", type="string", example="+31 6 12345678"),
     *             @OA\Property(property="isSSO", type="boolean", example=false)
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="User registered successfully",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=true),
     *             @OA\Property(property="data", type="object")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error"
     *     )
     * )
     * Handle an incoming registration request.
     *
     * @throws \Illuminate\Validation\ValidationException
     */
    public function store(Request $request, GoogleSheetsService $sheetsService): JsonResponse
    {
        $isSSO = $request->boolean(key: 'isSSO') ?? false;

        if (!$isSSO) {
            $validator = Validator::make($request->all(), User::rules(), User::messages());

            if ($validator->fails()) {
                return ApiResponseClass::sendInvalidFields($validator->errors()->toArray(), User::messages());
            }
        }

        // TODO: move to background task
        $referral_code = '';

        do {
            $referral_code = Str::slug($request->name) . '-' . Str::random(6);
        } while (User::where('referral_code', $referral_code)->exists());

        try {
            User::create([
                ...$request->all(),
                'referral_code' => $referral_code
            ]);

            $sheetsService->exportModelToSpreadsheet(
                User::class,
                'Users'
            );
        } catch (\Exception $error) {
            Log::error($error->getMessage());

            if ($isSSO) {
                $sheetsService->exportModelToSpreadsheet(
                    User::class,
                    'Users'
                );
                return ApiResponseClass::sendSuccess(['user_created' => false]);
            } else {
                return ApiResponseClass::sendError();
            }
        }

        return ApiResponseClass::sendSuccess(['user_created' => true]);
    }
}